Cifar-10
A Convolutional Neural Network in the cifar-10 database. I was having some overfitting problems when I was coding it, but after I made an augumentation in the database, I got some better results. With two convolution layers in this code, I got a 71% accuracy in both training and testing bases, in 15 epochs.